separated out table and row and renamed them
