+++ front/src/components/spellRow/SpellRow.module.css
