+++ front/src/components/spellRow/types.tsx
@@ -1,5 +1,5 @@
 export type Spell = {
     name: string;
     description: string;
-    levels: number[];
+    level: number;
 };
