+++ front/src/components/spellTable/SpellTable.module.css
